MOAR SPEED
Rip out the fancy avoid-roundoff-errors calculations, since we're
dealing with arbitrary precision numerics anyway; by minimizing
the number of calls in the transition function we keep the per-row
overhead to a minimum.  By padding the precision of the n_prime
division in the final calculation, we should get a result which is
precise if rounded to the input precision, and to at least the
minimum numeric precision regardless.

On a -O0/casserts build this gets a bit more than 3x speedup from
the previous version, and gets us to within a factor of 2-3 of the
performance of builtin stddev_samp.